Add req origin to controllers. Health tests completed

+++ src/controllers/HealthController.js
@@ -10,6 +10,9 @@ const Health = require('../models/Health');
 module.exports = {
   async create(request, response) {
     const id = currentUserId(request);
+    logger.info(`Request origin: ${request.ip}`);
+    logger.info(`Inbound request to /health/create. User (${id})`);
+
     try {
       const user = await User.findById(id);
 
@@ -93,11 +96,13 @@ module.exports = {
         $push: { healthCheckpoints: healthCheckpoint._id },
       });
 
+      logger.info(`User (${id}) added a new health checkpoint successfully`);
       return response.json({
         statusCode: 200,
         message: 'New checkpoint successfully added',
       });
     } catch (exc) {
+      logger.info(`Exception occurred when trying to add Health Checkpoint to User (${id}). Details: ${exc}`);
       return response.status(500).json({
         statusCode: 500,
         message: `Error while saving new checkpoint to user(${id})`,
@@ -109,6 +114,8 @@ module.exports = {
   // Show method
   async show(request, response) {
     const id = currentUserId(request);
+    logger.info(`Request origin: ${request.ip}`);
+    logger.info(`Inbound request to /health/show. User (${id})`);
 
     let { startDate, endDate = new Date() } = request.body;
 
@@ -129,6 +136,7 @@ module.exports = {
         startDate = new Date(startDate);
         endDate = new Date(endDate);
 
+        logger.info(`Checkpoints between ${startDate} and ${endDate} returned to user (${id}).`);
         return response.json({
           statusCode: 200,
           startDate,
@@ -137,6 +145,7 @@ module.exports = {
         });
       }
 
+      logger.info(`All health checkpoints returned to user (${id}).`);
       return response.json({
         statusCode: 200,
         data: user.healthCheckpoints,
@@ -155,14 +164,29 @@ module.exports = {
   // Delete method
   async delete(request, response) {
     const { id } = request.params;
+    const userId = currentUserId(request);
+    logger.info(`Request origin: ${request.ip}`);
+    logger.info(`Inbound request to /health/delete. User (${id})`);
 
     try {
+      // Removes reference from User
+      const previousUser = await User.findOneAndUpdate({ _id: userId }, { $pullAll: { healthCheckpoints: [id] } });
+      if (!previousUser.healthCheckpoints.includes(id)) {
+        logger.warn(`User (${userId}) tried to delete a checkpoint (${id}) from someone else`);
+
+        return response.status(403).json({
+          statusCode: 403,
+          errorCode: errors.RESOURCE_OWNERSHIP_MISMATCH,
+          message: 'Current user is NOT allowed to access this resource',
+        });
+      }
+
       // Delete from Health collection
       const checkpoint = await Health.findByIdAndDelete(id);
-      // Removes reference from User
-      await User.findOneAndUpdate({ _id: currentUserId(request) }, { $pullAll: { healthCheckpoints: [id] } });
 
       if (checkpoint == null) {
+        logger.warn(`User (${userId}) tried to delete a unknown checkpoint (${id})`);
+
         return response.status(404).json({
           statusCode: 404,
           errorCode: errors.RESOURCE_NOT_IN_DATABASE,
@@ -170,6 +194,7 @@ module.exports = {
         });
       }
 
+      logger.info(`User (${userId}) deleted a checkpoint (${id}) successfully`);
       return response.json({
         statusCode: 200,
         data: checkpoint,
